requirements.txt
Django == 3.1
djangorestframework == 3.11.1
numpy == 1.19.1
pandas == 1.1.0
pyyaml == 5.3.1
autopep8 == 1.5.4
pycodestyle == 2.6.0
opentrons >= 3.16.1

aiohttp==3.4.4
aionotify==0.2.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
biopython==1.77
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
flake8==3.8.3
fuzzywuzzy==0.18.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
jsonschema==3.1.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
multidict==4.5.2
networkx==2.4
primers==0.2.6
pyflakes==2.2.0
pylint==2.5.3
pyrsistent==0.15.4
pySBOL >= 2.3.2
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.3
requests==2.24.0
seqfold==0.7.4
six==1.12.0
sqlparse==0.3.1
synbio==0.6.17
toml==0.10.1
urllib3==1.25.9
urwid==1.3.1
wrapt==1.12.1
yarl==1.3.0
zipp==0.6.0
